Fixed missing dependencies during runtime.
Added script to check dependencies used from the host,
which allowed to detect missing libraries linked during
build but not present in the AppImage.

To broaden the compatibility and usage of the file, using
more libraries where then bundled.

Also fixed a missing install directive into AppDir of the
LuaJIT custom build.